# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SIGAT API
  version: 1.0.0
  description: Contrato inicial da API do SIGAT, alinhado aos fluxos consumidos pelo frontend.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: API versionada atrás do Nginx
...
paths:
...
  /reports/inventory-summary:
    get:
      tags: [Reports]
      summary: Gera a prévia autenticada do relatório de inventário consolidado
      operationId: getInventorySummaryReport
      description: Usuários de Unidade recebem somente sua unidade autenticada. Administradores DITEL podem consultar o consolidado estadual ou filtrar por unidade ativa.
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/UnitId' }
        - { $ref: '#/components/parameters/Situation' }
        - { $ref: '#/components/parameters/ReportPeriod' }
      responses:
        '200':
          description: Relatório sem senhas, tokens, hashes ou identificadores de sessão
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryReportResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/inventory-summary/export:
    get:
      tags: [Reports]
      summary: Exporta o relatório de inventário consolidado como CSV ou PDF
      operationId: exportInventorySummaryReport
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/UnitId' }
        - { $ref: '#/components/parameters/Situation' }
        - { $ref: '#/components/parameters/ReportPeriod' }
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [csv, pdf] }
      responses:
        '200':
          description: Arquivo exportado dentro do escopo autenticado
          content:
            text/csv:
              schema: { type: string }
            application/pdf:
              schema: { type: string, format: binary }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sigat_session
  parameters:
...
    Situation:
      name: situation
      in: query
      schema:
        $ref: '#/components/schemas/EquipmentSituation'
    UnitId: { name: unitId, in: query, description: Aplicável somente ao administrador DITEL., schema: { type: string } }
    ReportPeriod:
      name: period
      in: query
      description: Período de referência para o relatório, baseado na data de cadastro do equipamento.
      schema:
        type: string
        enum: [july_2026, q2_2026]
...
  responses:
    BadRequest:
      description: Dados inválidos
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
...
    Unauthorized:
      description: Sessão ausente ou inválida
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Forbidden:
      description: Usuário autenticado sem uma unidade válida para acessar o dashboard da unidade
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            invalidUnitScope:
              value:
                code: FORBIDDEN
                message: Escopo de unidade obrigatório.
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: EQUIPMENT_NOT_FOUND }
        message: { type: string }
        details: {}
...
    UserRole:
      type: string
      enum: [ditel_admin, unit_user]
...
    UnitContext:
      type: object
      required: [id, name, acronym]
      properties:
        id: { type: string }
        name: { type: string }
        acronym: { type: string }
...
    EquipmentSituation:
      type: string
      enum: [active, maintenance, inactive, lost, written_off]
...
    InventoryReportCounters:
      type: object
      required: [total, active, maintenance, inactive, lost, writtenOff, attention]
      properties:
        total: { type: integer, minimum: 0 }
        active: { type: integer, minimum: 0 }
        maintenance: { type: integer, minimum: 0 }
        inactive: { type: integer, minimum: 0 }
        lost: { type: integer, minimum: 0 }
        writtenOff: { type: integer, minimum: 0 }
        attention: { type: integer, minimum: 0 }
    InventoryReportUnitSummary:
      allOf:
        - { $ref: '#/components/schemas/InventoryReportCounters' }
        - type: object
          required: [unit]
          properties:
            unit: { $ref: '#/components/schemas/UnitContext' }
    InventoryReportResponse:
      type: object
      required: [report, totals, units, generatedBy]
      properties:
        report:
          type: object
          required: [id, title, generatedAt, scope, filters]
          properties:
            id: { type: string, enum: [inventory-summary] }
            title: { type: string }
            generatedAt: { type: string, format: date-time }
            scope: { $ref: '#/components/schemas/UnitContext' }
            filters:
              type: object
              required: [situation]
              properties:
                situation:
                  oneOf:
                    - { $ref: '#/components/schemas/EquipmentSituation' }
                    - { type: 'null' }
        totals: { $ref: '#/components/schemas/InventoryReportCounters' }
        units:
          type: array
          items: { $ref: '#/components/schemas/InventoryReportUnitSummary' }
        generatedBy:
          type: object
          required: [name, role]
          properties:
            name: { type: string }
            role: { $ref: '#/components/schemas/UserRole' }
...